+++ common/datavalues/src/arrays/mod.rs
@@ -12,9 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#[cfg(test)]
-mod arithmetic_test;
-
 #[macro_use]
 mod arithmetic;
 mod builder;
