+++ common/datavalues/src/lib.rs
@@ -19,9 +19,6 @@
 #[macro_use]
 mod macros;
 
-#[cfg(test)]
-mod data_array_filter_test;
-
 #[allow(dead_code)]
 mod bit_util;
 mod data_array_filter;
