+++ common/datavalues/tests/it/arrays/ops/fill.rs
@@ -12,11 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use common_datavalues::prelude::*;
 use common_exception::Result;
-
-use crate::arrays::ops::fill::ArrayFull;
-use crate::arrays::ops::fill::ArrayFullNull;
-use crate::prelude::*;
 
 #[test]
 fn test_array_fill() -> Result<()> {
