+++ common/datavalues/tests/it/arrays/ops/group_hash.rs
@@ -12,10 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use common_datavalues::prelude::*;
 use common_exception::Result;
-
-use crate::arrays::ops::group_hash::GroupHash;
-use crate::prelude::*;
 
 #[test]
 fn test_group_hash() -> Result<()> {
