+++ common/datavalues/tests/it/arrays/ops/mod.rs
@@ -12,15 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use common_exception::Result;
-
-use crate::prelude::*;
-
-#[test]
-fn test_array_cast() -> Result<()> {
-    let array = DFUInt16Array::new_from_iter(1_u16..4u16);
-    let result = array.cast_with_type(&DataType::UInt8)?;
-    let expected = Series::new(vec![1_u8, 2, 3]);
-    assert!(result.series_equal(&expected));
-    Ok(())
-}
+mod agg;
+mod apply;
+mod cast;
+mod contain;
+mod fill;
+mod group_hash;
+mod r#if;
+mod scatter;
+mod take;
+mod take_random;
