+++ common/datavalues/tests/it/arrays/ops/scatter.rs
@@ -12,11 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use common_datavalues::arrays::get_list_builder;
+use common_datavalues::prelude::*;
 use common_exception::Result;
-
-use crate::arrays::get_list_builder;
-use crate::arrays::ops::scatter::ArrayScatter;
-use crate::prelude::*;
 
 #[test]
 fn test_scatter() -> Result<()> {
