+++ common/datavalues/tests/it/arrays/ops/take.rs
@@ -13,11 +13,9 @@
 // limitations under the License.
 
 //use common_arrow::arrow::array::ListArray;
+use common_datavalues::arrays::get_list_builder;
+use common_datavalues::prelude::*;
 use common_exception::Result;
-
-use super::TakeIdx;
-use crate::arrays::get_list_builder;
-use crate::prelude::*;
 
 #[test]
 fn test_take() -> Result<()> {
