+++ common/datavalues/tests/it/arrays/string.rs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::prelude::*;
+use common_datavalues::prelude::*;
 
 #[test]
 fn test_empty_array() {
