+++ common/datavalues/tests/it/data_array_filter.rs
@@ -12,10 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use common_datavalues::prelude::*;
 use common_exception::Result;
 use pretty_assertions::assert_eq;
-
-use crate::prelude::*;
 
 #[test]
 fn filter_batch_array() -> Result<()> {
