+++ common/datavalues/tests/it/series.rs
@@ -12,9 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use common_datavalues::prelude::*;
 use common_exception::Result;
-
-use crate::prelude::*;
 
 #[test]
 #[allow(clippy::eq_op)]
